GUI PYTHON: added script run_test_view_workspace.py
